Makign server a class
